discord_bot/requirements.txt
discord.py
requests
python-dotenv
Pillow
torch
transformers
playwright
pytesseract
configparser